requirements.txt
pycryptodome
loguru